% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar.R
\name{tr_models_anova_dbc}
\alias{tr_models_anova_dbc}
\title{ANOVA de um delineamento em blocos casualizados (DBC).}
\usage{
tr_models_anova_dbc(dados, resposta = "", tratamento = "", bloco = "")
}
\arguments{
\item{dados}{tabela.}

\item{resposta}{coluna da resposta.}

\item{tratamento}{coluna do tratamento (vira fator).}

\item{bloco}{coluna do bloco (vira fator).}
}
\value{
objeto `tr_models_fit`.
}
\description{
ANOVA de um delineamento em blocos casualizados (DBC).
}
