% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar.R
\name{tr_models_anova_dic}
\alias{tr_models_anova_dic}
\title{ANOVA de um delineamento inteiramente casualizado (DIC).}
\usage{
tr_models_anova_dic(dados, resposta = "", tratamento = "")
}
\arguments{
\item{dados}{tabela.}

\item{resposta}{coluna da resposta.}

\item{tratamento}{coluna do tratamento (vira fator).}
}
\value{
objeto `tr_models_fit`.
}
\description{
ANOVA de um delineamento inteiramente casualizado (DIC).
}
