% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar.R
\name{tr_models_anova_dql}
\alias{tr_models_anova_dql}
\title{ANOVA de um delineamento em quadrado latino (DQL).}
\usage{
tr_models_anova_dql(
  dados,
  resposta = "",
  tratamento = "",
  linha = "",
  coluna = ""
)
}
\arguments{
\item{dados}{tabela.}

\item{resposta}{coluna da resposta.}

\item{tratamento}{coluna do tratamento (vira fator).}

\item{linha, coluna}{colunas das duas restrições (viram fator).}
}
\value{
objeto `tr_models_fit`.
}
\description{
ANOVA de um delineamento em quadrado latino (DQL).
}
