% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumir.R
\name{tr_models_anova_table}
\alias{tr_models_anova_table}
\title{Quadro da ANOVA.}
\usage{
tr_models_anova_table(modelo, tipo_sq = "I")
}
\arguments{
\item{modelo}{objeto `tr_models_fit`.}

\item{tipo_sq}{`"I"` (sequencial), `"II"` ou `"III"`.}
}
\value{
objeto `tr_models_effects`.
}
\description{
Quadro da ANOVA.
}
