% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrato.R
\name{tr_models_coefs.tr_models_gls}
\alias{tr_models_coefs.tr_models_gls}
\title{Coeficientes do GLS: t com n − p gl e intervalo do `nlme::intervals`.}
\usage{
\method{tr_models_coefs}{tr_models_gls}(
  x,
  exponenciar = FALSE,
  escala = "unidade",
  confianca = 0.95,
  ...
)
}
\description{
Coeficientes do GLS: t com n − p gl e intervalo do `nlme::intervals`.
}
