% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes.R
\name{tr_models_cor_test}
\alias{tr_models_cor_test}
\title{Teste de correlação.}
\usage{
tr_models_cor_test(dados, x = "", y = "", metodo = "pearson")
}
\arguments{
\item{dados}{tabela.}

\item{x, y}{as duas colunas numéricas.}

\item{metodo}{`"pearson"`, `"spearman"` ou `"kendall"`.}
}
\value{
objeto `tr_models_test`.
}
\description{
Teste de correlação.
}
