% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplos.R
\name{tr_models_example}
\alias{tr_models_example}
\title{Carrega um conjunto de exemplo.}
\usage{
tr_models_example(dataset = "PlantGrowth")
}
\arguments{
\item{dataset}{nome do conjunto (ver a ajuda do nó).}
}
\value{
tibble.
}
\description{
Carrega um conjunto de exemplo.
}
