% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes.R
\name{tr_models_fisher_exact}
\alias{tr_models_fisher_exact}
\title{Exato de Fisher.}
\usage{
tr_models_fisher_exact(dados, linha = "", coluna = "")
}
\arguments{
\item{dados}{tabela.}

\item{linha, coluna}{as duas colunas categóricas.}
}
\value{
objeto `tr_models_test`.
}
\description{
Exato de Fisher.
}
