% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes.R
\name{tr_models_kruskal}
\alias{tr_models_kruskal}
\title{Kruskal-Wallis: dois ou mais grupos.}
\usage{
tr_models_kruskal(dados, resposta = "", grupo = "")
}
\arguments{
\item{dados}{tabela.}

\item{resposta}{coluna numérica.}

\item{grupo}{coluna dos grupos.}
}
\value{
objeto `tr_models_test`.
}
\description{
Kruskal-Wallis: dois ou mais grupos.
}
