% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naolinear.R
\name{tr_models_nls}
\alias{tr_models_nls}
\title{Regressão não linear com modelos prontos.}
\usage{
tr_models_nls(dados, resposta = "", preditor = "", modelo = "logístico")
}
\arguments{
\item{dados}{tabela.}

\item{resposta}{coluna da resposta.}

\item{preditor}{a coluna numérica da preditora (uma só).}

\item{modelo}{`"logístico"`, `"Michaelis-Menten"`, `"exponencial
assintótico"`, `"Gompertz"` ou `"linear-platô"`.}
}
\value{
objeto `tr_models_fit` de classe `tr_models_nls`.
}
\description{
Regressão não linear com modelos prontos.
}
