% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medias.R
\name{tr_models_plot_means}
\alias{tr_models_plot_means}
\title{Médias com intervalo de confiança e letras.}
\usage{
tr_models_plot_means(
  medias,
  letras = TRUE,
  aspecto = "16:9",
  tema = "padrão",
  titulo = "",
  rotulo_x = "",
  rotulo_y = "",
  legenda = "direita"
)
}
\arguments{
\item{medias}{objeto `tr_models_emm`.}

\item{letras}{escrever as letras acima dos intervalos.}
}
\description{
Médias com intervalo de confiança e letras.
}
