% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumir.R
\name{tr_models_random_effects}
\alias{tr_models_random_effects}
\title{Componentes de variância.}
\usage{
tr_models_random_effects(modelo)
}
\arguments{
\item{modelo}{objeto `tr_models_fit` de um misto (ou de parcela subdividida).}
}
\value{
tibble `grupo`, `componente`, `variancia`, `desvio_padrao`,
  `correlacao`, `proporcao`.
}
\description{
Componentes de variância.
}
