% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumir.R
\name{tr_models_random_test}
\alias{tr_models_random_test}
\title{Teste dos efeitos aleatórios: cada um sai, e o ajuste piora?}
\usage{
tr_models_random_test(modelo)
}
\arguments{
\item{modelo}{objeto `tr_models_fit` de um misto.}
}
\value{
objeto `tr_models_effects`.
}
\description{
Teste dos efeitos aleatórios: cada um sai, e o ajuste piora?
}
