% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose.R
\name{tr_models_resid.tr_models_dose}
\alias{tr_models_resid.tr_models_dose}
\title{A tabela da ANOVA com a curva ao lado: o resíduo é o desvio de cada parcela
à curva (bloco e falta de ajuste inclusos), padronizado pelo erro da ANOVA.}
\usage{
\method{tr_models_resid}{tr_models_dose}(x)
}
\description{
A tabela da ANOVA com a curva ao lado: o resíduo é o desvio de cada parcela
à curva (bloco e falta de ajuste inclusos), padronizado pelo erro da ANOVA.
}
