% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrato.R
\name{tr_models_resid.tr_models_gls}
\alias{tr_models_resid.tr_models_gls}
\title{GLS: resíduos normalizados (descontadas a correlação e a variância do modelo).}
\usage{
\method{tr_models_resid}{tr_models_gls}(x)
}
\description{
GLS: resíduos normalizados (descontadas a correlação e a variância do modelo).
}
