% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes.R
\name{tr_models_shapiro}
\alias{tr_models_shapiro}
\title{Shapiro-Wilk numa coluna.}
\usage{
tr_models_shapiro(dados, variavel = "")
}
\arguments{
\item{dados}{tabela.}

\item{variavel}{coluna numérica testada.}
}
\value{
objeto `tr_models_test`.
}
\description{
Shapiro-Wilk numa coluna.
}
