% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressupostos.R
\name{tr_models_shapiro_residuals}
\alias{tr_models_shapiro_residuals}
\title{Shapiro-Wilk nos resíduos do modelo.}
\usage{
tr_models_shapiro_residuals(modelo)
}
\arguments{
\item{modelo}{objeto `tr_models_fit`.}
}
\value{
objeto `tr_models_test`.
}
\description{
Shapiro-Wilk nos resíduos do modelo.
}
