% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrato.R
\name{tr_models_stats.tr_models_gls}
\alias{tr_models_stats.tr_models_gls}
\title{GLS: sigma do `nlme` e gl do resíduo n − p (os dos coeficientes e do quadro).}
\usage{
\method{tr_models_stats}{tr_models_gls}(x)
}
\description{
GLS: sigma do `nlme` e gl do resíduo n − p (os dos coeficientes e do quadro).
}
