% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agricolae.R
\name{tr_models_waller_duncan}
\alias{tr_models_waller_duncan}
\title{Teste de Waller-Duncan (razão K).}
\usage{
tr_models_waller_duncan(modelo, tratamento = "", k = 100L)
}
\arguments{
\item{modelo}{objeto `tr_models_fit` de uma ANOVA.}

\item{tratamento}{fator (ou até 3, para as combinações).}

\item{k}{razão de custos dos erros tipo I e II: 100 (~5%), 500 (~1%), 50 (~10%).}
}
\value{
objeto `tr_models_emm`.
}
\description{
Teste de Waller-Duncan (razão K).
}
